Clean up markdown data set
I was originally planning on creating an iterator to see how many hastags
are in the markdown to determine the h level, but that's over engineering it.
h1-h6 has and will always be a conding standard, so we can manually declare them.

Also converted data array to associative, since the key vales will be super helpful.